+++ app/src/main/java/com/example/ppjoke/ui/detail/FeedVideoDetailActivity.kt
@@ -79,7 +79,7 @@ class FeedVideoDetailActivity :
             myActivityLauncher.launch(intent)
         }
         binding.fullscreenAuthorInfo.btnFollow.setOnClickListener {
-            val isFollow= InteractionPresenter.toggleFollowUser(feed?.author?.userId!!)
+            val isFollow= InteractionPresenter.toggleFollowUser(feed?.author?.userId!!, this)
             mViewModel!!.userRelation.value=isFollow
         }
         binding.authorInfo.authorAvatar.setOnClickListener(this)
@@ -137,21 +137,23 @@ class FeedVideoDetailActivity :
                 adapter!!.setOnItemChildClickListener { _, view, position ->
                     when (view.id) {
                         R.id.comment_like -> {
-                            mViewModel?.commentLike(adapter!!.data[position].commentId!!)
-                            println("喜欢评论：${mViewModel!!.isLike}")
-                            adapter!!.data[position].ugc!!.hasLiked=mViewModel!!.isLike
-                            if (mViewModel!!.isLike) {
-                                adapter!!.data[position].ugc!!.likeCount =
-                                    adapter!!.data[position].ugc!!.likeCount!!.plus(
-                                        1
-                                    )
-                            } else {
-                                adapter!!.data[position].ugc!!.likeCount =
-                                    adapter!!.data[position].ugc!!.likeCount?.minus(
-                                        1
-                                    )
+                            if(InteractionPresenter.checkIsLogin(this, mViewModel!!.userId)) {
+                                mViewModel?.commentLike(adapter!!.data[position].commentId!!)
+                                adapter!!.data[position].ugc!!.hasLiked = mViewModel!!.isLike
+                                if (mViewModel!!.isLike) {
+                                    adapter!!.data[position].ugc!!.likeCount =
+                                        adapter!!.data[position].ugc!!.likeCount!!.plus(
+                                            1
+                                        )
+                                } else {
+                                    adapter!!.data[position].ugc!!.likeCount =
+                                        adapter!!.data[position].ugc!!.likeCount?.minus(
+                                            1
+                                        )
+                                }
+                                adapter!!.notifyItemChanged(position, "commentAdd")
+
                             }
-                            adapter!!.notifyItemChanged(position, "commentAdd")
                         }
                         R.id.comment_author_avatar -> {
                             val intent = Intent(this, ProfileActivity::class.java)
@@ -169,15 +171,17 @@ class FeedVideoDetailActivity :
                                 .isDestroyOnDismiss(true)
                                 .dismissOnTouchOutside(false)
                                 .asConfirm("提示","确定要删除该条评论？") {
-                                    mViewModel!!.deleteComment(
-                                        adapter!!.data[position].itemId!!,
-                                        adapter!!.data[position].commentId!!
-                                    ) { isSuccess ->
-                                        if (isSuccess) {
-                                            adapter!!.data.removeAt(position)
-                                            adapter!!.notifyItemRemoved(position)
-                                        }else{
-                                            toastShort("出错了，删除失败！")
+                                    if(InteractionPresenter.checkIsLogin(this, mViewModel!!.userId)) {
+                                        mViewModel!!.deleteComment(
+                                            adapter!!.data[position].itemId!!,
+                                            adapter!!.data[position].commentId!!
+                                        ) { isSuccess ->
+                                            if (isSuccess) {
+                                                adapter!!.data.removeAt(position)
+                                                adapter!!.notifyItemRemoved(position)
+                                            } else {
+                                                toastShort("出错了，删除失败！")
+                                            }
                                         }
                                     }
 
@@ -225,10 +229,10 @@ class FeedVideoDetailActivity :
     override fun onClick(v: View?) {
         when(v?.id){
             R.id.btn_like->{
-                InteractionPresenter.toggleFeedLikeInternal(feed!!)
+                InteractionPresenter.toggleFeedLikeInternal(feed!!,this)
             }
             R.id.btn_collect->{
-                InteractionPresenter.toggleFeedFeedFavorite(feed!!)
+                InteractionPresenter.toggleFeedFeedFavorite(feed!!,this)
             }
             R.id.btn_share->{
                 InteractionPresenter.openShare(this,feed!!)
@@ -244,7 +248,7 @@ class FeedVideoDetailActivity :
                 myActivityLauncher.launch(intent)
             }
             R.id.btn_follow->{
-                val isFollow= InteractionPresenter.toggleFollowUser(feed?.author?.userId!!)
+                val isFollow= InteractionPresenter.toggleFollowUser(feed?.author?.userId!!,this)
                 mViewModel!!.userRelation.value=isFollow
             }
         }
